requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.6.4
requests==2.31.0
networkx==3.2.1
groq==1.1.2
google-generativeai==0.8.6
python-dotenv==1.0.1
apprise==1.8.0
pytest==8.1.1
pytest-asyncio==0.23.6
